+++ packages/tooling/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @richie-router/tooling
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @richie-router/core@0.1.4
+
 ## 0.1.4
 
 ### Patch Changes